Move kalevala specific files into its own folder
