+++ example/assets/js/client.js
@@ -1,23 +1,12 @@
 import React from 'react';
 import { Provider } from 'react-redux';
-import { compose, createStore } from 'redux';
 
-import { Creators } from "./redux/actions";
-import { kalevalaReducer } from "./redux";
+import { Prompt, SocketProvider, Terminal } from "./kalevala/components";
 
-import { Prompt, SocketProvider, Terminal } from "./components";
+import { makeStore } from "./store.js";
 
 let body = document.getElementById("body");
 
-const makeStore = () => {
-  const composeEnhancers =
-    typeof window === 'object' && window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ ?
-      window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__({}) : compose;
-
-  const enhancer = composeEnhancers();
-
-  return createStore(kalevalaReducer, enhancer);
-}
 
 let store = makeStore();
