+++ example/assets/js/kalevala/components/Prompt.js
