+++ example/assets/js/kalevala/components/SocketProvider.js
