+++ example/assets/js/kalevala/components/Terminal.js
