+++ example/assets/js/kalevala/components/colors.js
