+++ example/assets/js/kalevala/components/index.js
