+++ example/assets/js/kalevala/redux/actions.js
