+++ example/assets/js/kalevala/redux/index.js
