+++ example/assets/js/kalevala/redux/promptReducer.js
