+++ example/assets/js/kalevala/redux/promptReducer.test.js
