+++ example/assets/js/kalevala/redux/socketReducer.js
