+++ example/assets/js/kalevala/redux/socketReducer.test.js
