+++ example/assets/js/kalevala/socket.js
